requirements.txt
pydantic
google-generativeai
firecrawl-py
ddgs
python-dotenv
groq
langgraph
langchain
langsmith
tenacity